@@ -2,19 +2,21 @@
   "name": "client",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:3001",
   "dependencies": {
-    "@apollo/client": "^3.5.8",
+    "@apollo/client": "^3.7.17",
+    "@apollo/react-hooks": "^4.0.0",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.14.3",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "bootstrap": "^5.2.3",
-    "jwt-decode": "^2.2.0",
-    "graphql": "^15.4.0",
-    "react": "^16.13.1",
-    "react-bootstrap": "^2.7.0",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2",
+    "graphql-tag": "^2.12.6",
+    "jwt-decode": "^3.1.2",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.8.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.14.2",
     "react-scripts": "^5.0.1"
   },
   "scripts": {
