@@ -1,16 +1,79 @@
 import React, { useState, useEffect } from 'react';
-import {
-  Container,
-  Col,
-  Form,
-  Button,
-  Card,
-  Row
-} from 'react-bootstrap';
+import { Form, Button, Container, Row, Col, Card } from 'react-bootstrap';
+// import { saveBook, searchGoogleBooks } from '../utils/API';
+// import { saveBookIds, getSavedBookIds } from '../utils/localStorage';
+import { useQuery, useMutation } from '@apollo/client';
+import { SAVE_BOOK } from '../utils/mutations';
+import { QUERY_BOOK } from '../utils/queries';
+import AuthContext from '../context/authContext';
 
-import Auth from '../utils/auth';
-import { saveBook, searchGoogleBooks } from '../utils/API';
-import { saveBookIds, getSavedBookIds } from '../utils/localStorage';
+// SearchForm component
+const SearchForm = ({ searchInput, setSearchInput, handleFormSubmit }) => {
+  return (
+    <Form onSubmit={handleFormSubmit}>
+      <Row>
+        <Col xs={12} md={8}>
+          <Form.Control
+            name='searchInput'
+            value={searchInput}
+            onChange={(e) => setSearchInput(e.target.value)}
+            type='text'
+            size='lg'
+            placeholder='Search for a book'
+          />
+        </Col>
+        <Col xs={12} md={4}>
+          <Button type='submit' variant='success' size='lg'>
+            Submit Search
+          </Button>
+        </Col>
+      </Row>
+    </Form>
+  );
+};
+
+// BookList component
+const BookList = ({ searchedBooks, handleSaveBook, savedBookIds }) => {
+  // create state to hold saved bookId values
+  return (
+    <Container>
+      <h2 className='pt-5'>
+        {searchedBooks.length
+          ? `Viewing ${searchedBooks.length} results:`
+          : 'Search for a book to begin'}
+      </h2>
+      <Row>
+        {searchedBooks.map((book) => {
+          return (
+            <Col md="4">
+              <Card key={book.bookId} border='dark'>
+                {book.image ? (
+                  <Card.Img src={book.image} alt={`The cover for ${book.title}`} variant='top' />
+                ) : null}
+                <Card.Body>
+                  <Card.Title>{book.title}</Card.Title>
+                  <p className='small'>Authors: {book.authors}</p>
+                  <Card.Text>{book.description}</Card.Text>
+                  {Auth.loggedIn() && (
+                    <Button
+                      disabled={savedBookIds?.some((savedBookId) => savedBookId === book.bookId)}
+                      className='btn-block btn-info'
+                      onClick={() => handleSaveBook(book.bookId)}>
+                      {savedBookIds?.some((savedBookId) => savedBookId === book.bookId)
+                        ? 'This book has already been saved!'
+                        : 'Save this Book!'}
+                    </Button>
+                  )}
+                </Card.Body>
+              </Card>
+            </Col>
+          );
+        })}
+      </Row>
+    </Container>
+  )
+};
+
 
 const SearchBooks = () => {
   // create state for holding returned google api data
@@ -18,8 +81,11 @@ const SearchBooks = () => {
   // create state for holding our search field data
   const [searchInput, setSearchInput] = useState('');
 
-  // create state to hold saved bookId values
   const [savedBookIds, setSavedBookIds] = useState(getSavedBookIds());
+
+  const [loading, setLoading] = useState(false);
+
+  const [error, setError] = useState('');
 
   // set up useEffect hook to save `savedBookIds` list to localStorage on component unmount
   // learn more here: https://reactjs.org/docs/hooks-effect.html#effects-with-cleanup
@@ -32,8 +98,16 @@ const SearchBooks = () => {
     event.preventDefault();
 
     if (!searchInput) {
+      setError('Please enter a search term!')
       return false;
     }
+
+    if (!/^[a-xA-Z0-9 ]*$/.test(searchInput)) {
+      setError('Please enter a valid search term!')
+      return false;
+    }
+
+    setLoading(true);
 
     try {
       const response = await searchGoogleBooks(searchInput);
@@ -54,8 +128,14 @@ const SearchBooks = () => {
 
       setSearchedBooks(bookData);
       setSearchInput('');
+
+      setLoading(false);
+
     } catch (err) {
+      setError(err.message);
       console.error(err);
+
+      setLoading(false);
     }
   };
 
@@ -81,72 +161,31 @@ const SearchBooks = () => {
       // if book successfully saves to user's account, save book id to state
       setSavedBookIds([...savedBookIds, bookToSave.bookId]);
     } catch (err) {
+      setError(err.message);
       console.error(err);
     }
   };
 
   return (
     <>
-      <div className='text-light bg-dark pt-5'>
-        <Container>
-          <h1>Search for Books!</h1>
-          <Form onSubmit={handleFormSubmit}>
-            <Row>
-              <Col xs={12} md={8}>
-                <Form.Control
-                  name='searchInput'
-                  value={searchInput}
-                  onChange={(e) => setSearchInput(e.target.value)}
-                  type='text'
-                  size='lg'
-                  placeholder='Search for a book'
-                />
-              </Col>
-              <Col xs={12} md={4}>
-                <Button type='submit' variant='success' size='lg'>
-                  Submit Search
-                </Button>
-              </Col>
-            </Row>
-          </Form>
-        </Container>
-      </div>
 
-      <Container>
-        <h2 className='pt-5'>
-          {searchedBooks.length
-            ? `Viewing ${searchedBooks.length} results:`
-            : 'Search for a book to begin'}
-        </h2>
-        <Row>
-          {searchedBooks.map((book) => {
-            return (
-              <Col md="4">
-                <Card key={book.bookId} border='dark'>
-                  {book.image ? (
-                    <Card.Img src={book.image} alt={`The cover for ${book.title}`} variant='top' />
-                  ) : null}
-                  <Card.Body>
-                    <Card.Title>{book.title}</Card.Title>
-                    <p className='small'>Authors: {book.authors}</p>
-                    <Card.Text>{book.description}</Card.Text>
-                    {Auth.loggedIn() && (
-                      <Button
-                        disabled={savedBookIds?.some((savedBookId) => savedBookId === book.bookId)}
-                        className='btn-block btn-info'
-                        onClick={() => handleSaveBook(book.bookId)}>
-                        {savedBookIds?.some((savedBookId) => savedBookId === book.bookId)
-                          ? 'This book has already been saved!'
-                          : 'Save this Book!'}
-                      </Button>
-                    )}
-                  </Card.Body>
-                </Card>
-              </Col>
-            );
-          })}
-        </Row>
-      </Container>
+      {/* Display error if there is one */}
+      {error && <div className="alert alert-danger mb-3">{error}</div>}
+
+      {/* Display loading message when there is a request in progress */}
+      {loading && <p>Loading results...</p>}
+
+      <SearchForm
+        searchInput={searchInput}
+        setSearchInput={setSearchInput}
+        handleFormSubmit={handleFormSubmit}
+      />
+
+      <BookList books={searchedBooks}
+        handleSaveBook={handleSaveBook}
+        setSavedBookIds={setSavedBookIds}
+      />
+
     </>
   );
 };
