@@ -2,15 +2,14 @@ import { gql } from '@apollo/client';
 
 // Mutations for login, create user, save book, and remove book
 export const LOGIN_USER = gql`
-    mutation loginUser($email: String!, $password: String!) {
-        login(email: $email, password: $password) {
+    mutation loginUser($loginInput: LoginInput!) {
+        loginUser(loginInput: $loginInput) {
+            email
+            username
             token
-            user {
-                _id
-                username
-            }
+
+                }
         }
-    }
 `;
 
 export const CREATE_USER = gql`
