@@ -2,7 +2,7 @@ import { gql } from '@apollo/client';
 
 // Query to fetch logged-in user's data (including saved books)
 export const QUERY_ME = gql`
-    query me {
+    query user {
         _id
         username
         email
@@ -15,5 +15,17 @@ export const QUERY_ME = gql`
             image
             link
         }
+    }
+`;
+
+// Query to fetch books from Google Books API
+export const QUERY_BOOK = gql`
+    query book {
+        bookId
+        authors
+        description
+        title
+        image
+        link
     }
 `;
