@@ -1,57 +1,37 @@
 const { Schema, model } = require('mongoose');
-const bcrypt = require('bcrypt');
 
-// import schema from Book.js
-const { Book } = require('./Book');
-
-const userSchema = new Schema(
-  {
+const userSchema = new Schema({
+  
     username: {
       type: String,
       required: true,
       unique: true,
+      trim: true,
+      minlength: 3
     },
+
     email: {
       type: String,
       required: true,
       unique: true,
       match: [/.+@.+\..+/, 'Must use a valid email address'],
+      trim: true
     },
+
     password: {
       type: String,
       required: true,
+      trim: true,
+      minlength: 5
     },
-    // set savedBooks to be an array of data that adheres to the bookSchema
-    savedBooks: [Book],
+
+    token: {
+      type: String,
+      required: true,
+      unique: true,
+      trim: true
   },
-  // set this to use virtual below
-  {
-    toJSON: {
-      virtuals: true,
-    },
-  }
-);
 
-// hash user password
-userSchema.pre('save', async function (next) {
-  if (this.isNew || this.isModified('password')) {
-    const saltRounds = 10;
-    this.password = bcrypt.hash(this.password, saltRounds);
-  }
-
-  next();
 });
 
-// custom method to compare and validate password for logging in
-userSchema.methods.isCorrectPassword = async function (password) {
-  return bcrypt.compare(password, this.password);
-};
-
-// when we query a user, we'll also get another field called `bookCount` with the number of saved books we have
-userSchema.virtual('bookCount').get(function () {
-  return this.savedBooks.length;
-});
-
-const User = model('User', userSchema);
-
-module.exports = User;
+module.exports = model('User', userSchema);
