@@ -1,47 +1,99 @@
-const { User } = require('../models');
-const signToken = require('../utils/auth');
+const User = require('../models/User');
+const { ApolloError } = require('apollo-server-express');
+const Book = require('../models/Book');
+const bcrypt = require('bcrypt');
+const jwt = require('jsonwebtoken');
+const fetch = require('node-fetch');
 
-const resolvers = {
+module.exports = {
     Query: {
         // Query single user by id or username
-        user: async (_, { _id, username }) => {
-            return User.findOne({
+        user: async (_, { id, username }) => {
+            User.findOne({
                 $or: [{ _id: id }, { username }],
             });
         },
+
+        // Query books from Google Books API
+        book: async (_, { query }) => {
+            const response = await fetch(`https://www.googleapis.com/books/v1/volumes?q=${query}`);
+            const res = await response.json();
+
+            if(!res) {
+                throw new Error('No books found!');
+            }
+
+            const bookData = res.items.map(({ book }) => ({
+                bookId: book.id,
+                authors: book.volumeInfo.authors,
+                description: book.volumeInfo.description,
+                title: book.volumeInfo.title,
+                image: book.volumeInfo.imageLinks?.thumbnail,
+                link: book.volumeInfo.infoLink,
+            }));
+
+            await Book.insertMany(bookData);
+            return bookData;
+        }
     },
     Mutation: {
-        // Create new user with signed token and send it to client
-        createUser: async (_, { input }) => {
-            const user = await User.create(input);
+        // Register new user with signed token and send it to client
+        registerUser: async (_, { registerInput: { username, email, password } }) => {
 
-            if(!user) {
-                throw new Error('Something went wrong!');
+            // Check if user already exists
+            const oldUser = await User.findOne({ email, username });
+            if (oldUser === email || oldUser === username) {
+                throw new ApolloError('User already exists with that email or username!');
             }
 
-            const token = signToken(user);
+            // Encrypt user's password
+            const encryptedPassword = await bcrypt.hash(password, 10);
 
-            return { token, user };
+            // Create the new user in the database
+            const newUser = new User({
+                username: username,
+                email: email.toLowerCase(),
+                password: encryptedPassword,
+            });
 
+            // Create the jwt and attach to user
+            const token = jwt.sign(
+                { user_id: newUser._id, email },
+                "dontputthesecrettokenhereinquotes",
+                { expiresIn: "2h" }
+            );
+
+            newUser.token = token;
+
+            // Save the user to the database
+            const res = await newUser.save();
+
+            return { id: res.id, ...res._doc };
         },
 
+
         // Login user with email and password and send token to client
-        login: async (_, { input }) => {
-            const user = await User.findOne({ $or: [{ username: input.username }, { email: input.email }] });
+        loginUser: async (_, { loginInput: { email, password } }) => {
 
-            if(!user) {
-                throw new Error('Cannot find this user!');
+            // Check for user's email in database
+            const user = await User.findOne({ email });
+
+            // If user exists, check password and create jwt
+            if (user && (await bcrypt.compare(password, user.password))) {
+
+                const token = jwt.sign(
+                    { user_id: user._id, email },
+                    "dontputthesecrettokenhereinquotes",
+                    { expiresIn: "2h" }
+                );
+
+                user.token = token;
+
+                return { id: user.id, ...user._doc };
+                // Else throw error if user doesn't exist or password is incorrect
+            } else {
+                throw new ApolloError("Incorrect email or password!");
             }
-
-            const correctPw = await user.isCorrectPassword(input.password);
-
-            if(!correctPw) {
-                throw new Error('Wrong password!');
-            }
-
-            const token = signToken(user);
-
-            return { token, user };
         },
 
         // Save book to user's `savedBooks` field by adding it to the set (to prevent duplicates)
@@ -75,5 +127,3 @@ const resolvers = {
         },
     },
 };
-
-module.exports = resolvers;
