@@ -1,11 +1,12 @@
 const { gql } = require('apollo-server-express');
 
 // Create GraphQL typeDefs
-const typeDefs = gql`
+module.exports = gql`
     type User {
         _id: ID!
         username: String!
         email: String!
+        token: String!
         bookCount: Int
         savedBooks: [Book]
     }
@@ -19,21 +20,16 @@ const typeDefs = gql`
         link: String
     }
 
-    type Auth {
-        token: ID!
-        user: User
-    }
-
     input LoginInput {
         email: String!
-        username: String!
         password: String!
     }
 
-    input CreateUserInput {
-        username: String!
+    input RegisterInput {
         email: String!
+        username: String!
         password: String!
+        confirmPassword: String!
     }
 
     input SaveBookInput {
@@ -46,16 +42,14 @@ const typeDefs = gql`
     }
 
     type Query {
-        me: User
+        book(query: String!): [Book!]!
+        user(id: ID!): User!
     }
 
     type Mutation {
-        login(input: LoginInput!): Auth
-        createUser(input: CreateUserInput!): Auth
+        loginUser(loginInput: LoginInput): User!
+        registerUser(registerInput: RegisterInput): User!
         saveBook(userId: ID!, book: SaveBookInput!): User
         removeBook(userId: ID!, bookId: String!): User
     }
 `;
-
-// Export typeDefs
-module.exports = typeDefs;
