@@ -1,36 +1,18 @@
+const { AuthenticationError } = require('apollo-server-express');
+
 const jwt = require('jsonwebtoken');
 
-// set token secret and expiration date
-const secret = 'mysecretsshhhhh';
-const expiration = '2h';
-
-module.exports = {
-  // function for our authenticated routes
-  authMiddleware: function ({ req }) {
-    // allows token to be sent via  req.query or headers
-    let token = req.query.token || req.headers.authorization;
-
-    // Separate "Bearer" from "<tokenvalue>
-    if (req.headers.authorization) {
-      token = token.split(' ').pop().trim();
+module.exports = (context) => {
+  const authHeader = context.req.headers.authorization;
+  if (authHeader) {
+    const token = authHeader.split('Bearer')[1];
+    if (token) {
+      try {
+        const user = jwt.verify(token, "dontputthesecrettokenhereinquotes");
+        return user;
+      } catch (error) {
+        throw new AuthenticationError('Invalid/Expired token');
+      }
     }
-
-    // if no token, return request object as is
-    if (!token) {
-      return req;
-    }
-
-    // Try to decode and attach user data to request object
-    try {
-      const { data } = jwt.verify(token, secret, { maxAge: expiration });
-      req.user = data;
-    } catch {
-      console.log('Invalid token');
-    }
-  },
-  signToken: function ({ username, email, _id }) {
-    const payload = { username, email, _id };
-
-    return jwt.sign({ data: payload }, secret, { expiresIn: expiration });
-  },
+  }
 };
